refactor(manifest): move configfile helper to cmdhelpers, migrate to slog
  Move cmd/cli/helper → cmdhelpers/configfile so linter and manifest
  commands share one package without cross-importing.

  Replace log.Logger/sliceWriter with slog.Handler in manifest CLI;
  captures messages without timestamp noise.

  GenerateDefinition now returns error on invalid flag config instead
  of silently logging and continuing.

  FlagDefinition switched from model/dto to modules/core/dto for
  consistency with the rest of the manifest pipeline.
