+++ cmd/cli/generate/manifest/manifest.go
@@ -1,13 +1,15 @@
 package manifest
 
 import (
+	"context"
 	"encoding/json"
 	"fmt"
-	"log"
+	"log/slog"
 	"os"
 	"strings"
 
 	"github.com/thomaspoignant/go-feature-flag/cmd/cli/helper"
+	configHelper "github.com/thomaspoignant/go-feature-flag/cmdhelpers/configfile"
 	"github.com/thomaspoignant/go-feature-flag/cmdhelpers/manifest"
 	"github.com/thomaspoignant/go-feature-flag/cmdhelpers/manifest/model"
 	"github.com/thomaspoignant/go-feature-flag/model/dto"
@@ -20,10 +22,10 @@ func NewManifest(configFile, configFormat, flagManifestDestination string) (Mani
 	if flagManifestDestination == "" {
 		return Manifest{}, fmt.Errorf("--flag_manifest_destination is mandatory")
 	}
-	flagDTOs, err := helper.LoadConfigFile(
+	flagDTOs, err := configHelper.LoadConfigFile(
 		configFile,
 		configFormat,
-		helper.ConfigFileDefaultLocations,
+		configHelper.ConfigFileDefaultLocations,
 	)
 	if err != nil {
 		return Manifest{}, err
@@ -42,13 +44,16 @@ type Manifest struct {
 // Generate the manifest file
 func (m *Manifest) Generate() (helper.Output, error) {
 	var logs []string
-	logx := fflog.ConvertToFFLogger(log.New(&sliceWriter{logs: &logs}, "", log.LstdFlags))
+	handler := slogHandlerForLogs(&logs)
+	logx := fflog.FFLogger{
+		LeveledLogger: slog.New(handler),
+	}
 	flags := make(map[string]flag.InternalFlag)
 	for k, v := range m.dtos {
 		flags[k] = dtoCore.ConvertDtoToInternalFlag(v)
 	}
 	output := helper.Output{}
-	definitions, err := manifest.GenerateDefinition(flags, *logx)
+	definitions, err := manifest.GenerateDefinition(flags, logx)
 	if len(logs) > 0 {
 		output.Add(strings.Join(logs, "\n"), helper.WarnLevel)
 	}
@@ -80,14 +85,24 @@ func (m *Manifest) storeManifest(definitionsJSON string) error {
 	return os.WriteFile(m.flagManifestDestination, []byte(definitionsJSON), 0600)
 }
 
-// sliceWriter is an io.Writer that captures each Write call as an entry
-// in the underlying string slice. It is used to redirect a *log.Logger's
-// output into an in-memory array instead of stdout/stderr.
-type sliceWriter struct {
+// captureHandler is a slog.Handler that appends each record's message to logs
+// (no timestamps/levels in the string; CLI applies WARNING/INFO via helper.Output).
+type captureHandler struct {
 	logs *[]string
 }
 
-func (w *sliceWriter) Write(p []byte) (int, error) {
-	*w.logs = append(*w.logs, strings.TrimRight(string(p), "\n"))
-	return len(p), nil
+func (h *captureHandler) Enabled(context.Context, slog.Level) bool { return true }
+
+func (h *captureHandler) Handle(_ context.Context, r slog.Record) error {
+	*h.logs = append(*h.logs, r.Message)
+	return nil
+}
+
+func (h *captureHandler) WithAttrs([]slog.Attr) slog.Handler { return h }
+
+func (h *captureHandler) WithGroup(string) slog.Handler { return h }
+
+// slogHandlerForLogs returns a slog.Handler that records log messages into logs.
+func slogHandlerForLogs(logs *[]string) slog.Handler {
+	return &captureHandler{logs: logs}
 }
