+++ cmd/cli/generate/manifest/testdata/output/flag-float-and-int.json
@@ -1,6 +1,32 @@
 {
+  "$schema": "https://raw.githubusercontent.com/open-feature/cli/refs/heads/main/schema/v0/flag-manifest.json",
   "flags": {
     "test-flag": {
+      "variations": {
+        "Default": 1.4,
+        "False": 2,
+        "True": 3
+      },
+      "targeting": [
+        {
+          "name": "rule1",
+          "query": "key eq \"random-key\"",
+          "percentage": {
+            "False": 0,
+            "True": 100
+          }
+        }
+      ],
+      "defaultRule": {
+        "name": "defaultRule",
+        "variation": "Default"
+      },
+      "experimentation": {},
+      "metadata": {
+        "defaultValue": 25,
+        "description": "this is a simple feature flag",
+        "issue-link": "https://jira.xxx/GOFF-01"
+      },
       "flagType": "float",
       "defaultValue": 25,
       "description": "this is a simple feature flag"
