+++ cmd/cli/generate/manifest/testdata/output/flag-no-default.json
@@ -1,11 +1,59 @@
 {
+  "$schema": "https://raw.githubusercontent.com/open-feature/cli/refs/heads/main/schema/v0/flag-manifest.json",
   "flags": {
     "test-flag2": {
+      "variations": {
+        "Default": false,
+        "False": false,
+        "True": true
+      },
+      "targeting": [
+        {
+          "name": "rule1",
+          "query": "key eq \"not-a-key\"",
+          "percentage": {
+            "False": 0,
+            "True": 100
+          }
+        }
+      ],
+      "defaultRule": {
+        "name": "defaultRule",
+        "variation": "Default"
+      },
+      "experimentation": {},
+      "metadata": {
+        "defaultValue": false
+      },
       "flagType": "boolean",
       "defaultValue": false,
       "description": ""
     },
     "test-flag3": {
+      "variations": {
+        "Default": 1,
+        "False": 2,
+        "True": 3
+      },
+      "targeting": [
+        {
+          "name": "rule1",
+          "query": "key eq \"not-a-key\"",
+          "percentage": {
+            "False": 0,
+            "True": 100
+          }
+        }
+      ],
+      "defaultRule": {
+        "name": "defaultRule",
+        "variation": "Default"
+      },
+      "experimentation": {},
+      "metadata": {
+        "defaultValue": 1,
+        "description": "totot"
+      },
       "flagType": "integer",
       "defaultValue": 1,
       "description": "totot"
