+++ cmd/lint/linter.go
@@ -3,7 +3,7 @@ package main
 import (
 	"fmt"
 
-	"github.com/thomaspoignant/go-feature-flag/cmd/cli/helper"
+	helper "github.com/thomaspoignant/go-feature-flag/cmdhelpers/configfile"
 )
 
 type Linter struct {
