+++ cmdhelpers/configfile/read_config_file.go
@@ -1,4 +1,4 @@
-package helper
+package configfile
 
 import (
 	"encoding/json"
