+++ cmdhelpers/configfile/testdata/invalid.json
